updated the batch, container

+++ resources/views/inventory/batch/addproduct.blade.php
@@ -26,6 +26,13 @@
     <?php echo displayAlert(); ?>
     <div class="page-header pt-3">
         <h3 class="page-title"> {{ $batch->name }} </h3>
+        <div class="page-inner">
+            @if($containers)
+                @foreach($containers as $key => $value)
+                    <input type="checkbox" name="filter_container" class = "form-control" id="filter_container" value="{{ $value->owner_name }}" />{{ $value->owner_name }}
+                @endforeach
+            @endif
+        </div>
         <nav aria-label="breadcrumb">
             <ol class="breadcrumb">
                 <li class="breadcrumb-item"><a href="{{ url('inventory/container/batch') }}">Batch</a></li>
@@ -86,45 +93,47 @@
                                                                 <tbody>
                                                                     @foreach ($allproducts as $prod)
                                                                         @if ($prod->category == $cus->id )
-                                                                            <tr id="row_{{ $tbl_inc }}" class="getRow">
-                                                                                <td>
-                                                                                    <label>{{ $prod->name }}</label>
-                                                                                    <input type="hidden" name="prodName[]"
-                                                                                        value="{{ $prod->id }}" />
-                                                                                    <input type="hidden" name="cat_id[]"
-                                                                                        value="{{ $prod->category }}" />
-                                                                                </td>
-                                                                                <td>
-                                                                                    <label style="font-weight: bolder;">{{ $prod->stock }}</label>
-                                                                                    <input type="hidden" name="initial_stock[]"
-                                                                                        class="iStock_{{ $tbl_inc }}"
-                                                                                        value="{{ $prod->stock }}" />
-                                                                                </td>
-                                                                                <td>
-                                                                                    <input type="hidden" class="form-control" name="cost[]" value="{{ $prod->price }}"/>
-                                                                                    <input type="text" readonly class="form-control stock_{{ $tbl_inc }}" name="stock[]" value="{{ $prod->stock }}" style="font-weight: bolder;" />
-                                                                                </td>
-                                                                                @foreach ($allmarks as $key => $mark)
+                                                                            @if($prod->stock > 0)
+                                                                                <tr id="row_{{ $tbl_inc }}" class="getRow">
                                                                                     <td>
-                                                                                        <input type="text" value="0"
-                                                                                            name="mark_{{ ($key + 1) + count($allmarks) * ($prod->id - 1) }}"
-                                                                                            class="form-control mkkk mark_{{ $tbl_inc }}" />
+                                                                                        <label>{{ $prod->name }}</label>
+                                                                                        <input type="hidden" name="prodName[]"
+                                                                                            value="{{ $prod->id }}" />
+                                                                                        <input type="hidden" name="cat_id[]"
+                                                                                            value="{{ $prod->category }}" />
                                                                                     </td>
-                                                                                    @php
-                                                                                        $td_inc++;
-                                                                                    @endphp
-                                                                                @endforeach
-                                                                                <td>
-                                                                                    <button type="button"
-                                                                                        onclick="deleteTblRow(this)"
-                                                                                        class="btn btn-danger btn-sm"><i
-                                                                                            class="fa fa-trash"></i></button>
-                                                                                </td>
-                                                                            </tr>
+                                                                                    <td>
+                                                                                        <label style="font-weight: bolder;">{{ $prod->stock }}</label>
+                                                                                        <input type="hidden" name="initial_stock[]"
+                                                                                            class="iStock_{{ $tbl_inc }}"
+                                                                                            value="{{ $prod->stock }}" />
+                                                                                    </td>
+                                                                                    <td>
+                                                                                        <input type="hidden" class="form-control" name="cost[]" value="{{ $prod->price }}"/>
+                                                                                        <input type="text" readonly class="form-control stock_{{ $tbl_inc }}" name="stock[]" value="{{ $prod->stock }}" style="font-weight: bolder;" />
+                                                                                    </td>
+                                                                                    @foreach ($allmarks as $key => $mark)
+                                                                                        <td>
+                                                                                            <input type="text" value="0"
+                                                                                                name="mark_{{ ($key + 1) + count($allmarks) * ($prod->id - 1) }}"
+                                                                                                class="form-control mkkk mark_{{ $tbl_inc }}" />
+                                                                                        </td>
+                                                                                        @php
+                                                                                            $td_inc++;
+                                                                                        @endphp
+                                                                                    @endforeach
+                                                                                    <td>
+                                                                                        <button type="button"
+                                                                                            onclick="deleteTblRow(this)"
+                                                                                            class="btn btn-danger btn-sm"><i
+                                                                                                class="fa fa-trash"></i></button>
+                                                                                    </td>
+                                                                                </tr>
 
-                                                                            @php
-                                                                                $tbl_inc++;
-                                                                            @endphp
+                                                                                @php
+                                                                                    $tbl_inc++;
+                                                                                @endphp
+                                                                            @endif
                                                                         @endif
                                                                     @endforeach
                                                                 </tbody>
